fix(change-risk): split tests.rs into tests/ to clear the god-file gate
sentrux gate: functions>25 && loc>400 (26 fns, 507 loc) on
change_risk/tests.rs. Split into tests/mod.rs (fixture helpers),
tests/predicates.rs (pure-logic + workflow-contract tests), and
tests/scoring.rs (git-fixture tests). Descendant-module visibility
keeps every helper and change_risk item reachable without new pub.
