+++ crates/code-intel-cli/src/change_risk/tests/scoring.rs
@@ -1,213 +1,8 @@
 use std::collections::BTreeSet;
-use std::path::{Path, PathBuf};
-use std::process::Command;
-use std::time::{SystemTime, UNIX_EPOCH};
 
-use super::git::{commits_in_range, diff_stats, resolve_repo_root_from, tip_token};
-use super::signals::{is_source_file, is_test_file, looks_like_fix_subject};
+use super::super::git::{commits_in_range, diff_stats, resolve_repo_root_from};
+use super::super::*;
 use super::*;
-
-/// A fresh, empty temp directory namespaced the same way `init_repo` names
-/// its repos, but without running `git init` — the starting point for both
-/// a to-be-initialized repo and (for the invalid-`--repo` test) a directory
-/// that must never look like one.
-fn temp_dir_for(name: &str) -> PathBuf {
-    let nonce = SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .unwrap()
-        .as_nanos();
-    let dir = std::env::temp_dir().join(format!(
-        "code-intel-change-risk-{name}-{nonce}-{}",
-        std::process::id()
-    ));
-    std::fs::create_dir_all(&dir).unwrap();
-    dir
-}
-
-fn init_repo(name: &str) -> PathBuf {
-    let repo = temp_dir_for(name);
-    for args in [
-        vec!["init", "--quiet"],
-        vec!["config", "user.name", "Change Risk Test"],
-        vec!["config", "user.email", "change-risk@example.invalid"],
-        vec!["config", "commit.gpgsign", "false"],
-    ] {
-        assert!(Command::new("git")
-            .args(args)
-            .current_dir(&repo)
-            .status()
-            .unwrap()
-            .success());
-    }
-    repo
-}
-
-fn write_file(repo: &Path, relative: &str, contents: &str) {
-    let path = repo.join(relative);
-    if let Some(parent) = path.parent() {
-        std::fs::create_dir_all(parent).unwrap();
-    }
-    std::fs::write(path, contents).unwrap();
-}
-
-fn commit(repo: &Path, message: &str) {
-    assert!(Command::new("git")
-        .args(["add", "."])
-        .current_dir(repo)
-        .status()
-        .unwrap()
-        .success());
-    assert!(Command::new("git")
-        .args(["commit", "--quiet", "-m", message])
-        .current_dir(repo)
-        .status()
-        .unwrap()
-        .success());
-}
-
-fn checkout_new_branch(repo: &Path, name: &str, start_point: Option<&str>) {
-    let mut args = vec!["checkout", "--quiet", "-b", name];
-    if let Some(start_point) = start_point {
-        args.push(start_point);
-    }
-    assert!(Command::new("git")
-        .args(args)
-        .current_dir(repo)
-        .status()
-        .unwrap()
-        .success());
-}
-
-fn rev_parse(repo: &Path, rev: &str) -> String {
-    let output = Command::new("git")
-        .args(["rev-parse", rev])
-        .current_dir(repo)
-        .output()
-        .unwrap();
-    assert!(output.status.success());
-    String::from_utf8_lossy(&output.stdout).trim().to_string()
-}
-
-#[test]
-fn weights_sum_to_one_hundred() {
-    assert_eq!(
-        WEIGHT_DIFF_SHAPE + WEIGHT_TEST_ASYMMETRY + WEIGHT_BUG_MAGNET + WEIGHT_CHURN,
-        100.0
-    );
-}
-
-#[test]
-fn is_source_file_matches_the_crates_src_glob() {
-    assert!(is_source_file("crates/code-intel-cli/src/main.rs"));
-    assert!(is_source_file("crates/code-intel-cli/src/sub/mod.rs"));
-    assert!(!is_source_file("crates/code-intel-cli/Cargo.toml"));
-    assert!(!is_source_file("crates/code-intel-cli/tests/it.rs"));
-    assert!(!is_source_file("README.md"));
-}
-
-#[test]
-fn is_test_file_matches_the_documented_heuristics() {
-    assert!(is_test_file("crates/code-intel-cli/tests/it.rs"));
-    assert!(is_test_file("crates/code-intel-cli/src/foo_test.rs"));
-    assert!(is_test_file("crates/code-intel-cli/src/test_foo.rs"));
-    assert!(!is_test_file("crates/code-intel-cli/src/main.rs"));
-}
-
-#[test]
-fn is_test_file_credits_this_crates_own_inline_test_module_convention() {
-    // Every multi-part module here keeps its tests in a sibling `tests.rs`.
-    // Before this was recognized, a PR adding one scored as "no tests
-    // touched" — the heaviest single signal in the formula.
-    for path in [
-        "crates/code-intel-cli/src/change_risk/tests.rs",
-        "crates/code-intel-cli/src/change_agenda/tests.rs",
-        "crates/code-intel-cli/src/file_gate/tests.rs",
-    ] {
-        assert!(is_test_file(path), "{path}");
-    }
-}
-
-#[test]
-fn is_test_file_does_not_credit_names_that_merely_contain_tests() {
-    // Whole-stem equality, not substring: these are ordinary source files.
-    for path in [
-        "crates/code-intel-cli/src/latest.rs",
-        "crates/code-intel-cli/src/contests.rs",
-        "crates/code-intel-cli/src/testsuite_helpers.rs",
-    ] {
-        assert!(!is_test_file(path), "{path}");
-    }
-}
-
-/// Contract-level guard on the emitted signal, not just the predicate that
-/// feeds it. The defect this fixes lived in what `testAsymmetry` reported,
-/// so a green `is_test_file` alone would not have caught it — and would not
-/// catch a future rewiring that leaves the predicate correct but stops
-/// consulting it.
-#[test]
-fn a_change_touching_an_inline_tests_module_is_not_scored_as_asymmetric() {
-    let files = vec![
-        (
-            40,
-            2,
-            "crates/code-intel-cli/src/change_risk/signals.rs".to_string(),
-        ),
-        (
-            30,
-            0,
-            "crates/code-intel-cli/src/change_risk/tests.rs".to_string(),
-        ),
-    ];
-    let scored = score_subset(&files, &FileHistory::new(), 1_700_000_000);
-
-    let asymmetry = &scored.signals["testAsymmetry"];
-    assert_eq!(asymmetry["testFilesChanged"].as_u64(), Some(1));
-    assert_eq!(asymmetry["asymmetric"], false);
-    assert_eq!(asymmetry["subscore"], 0.0);
-
-    let tests_row = scored
-        .files
-        .iter()
-        .find(|file| file["path"] == "crates/code-intel-cli/src/change_risk/tests.rs")
-        .expect("the tests.rs row must be reported");
-    assert_eq!(tests_row["isTestFile"], true);
-}
-
-/// The complementary half: the signal must still fire when a change really
-/// does touch source without touching any test. A fix that made every change
-/// look symmetric would pass the test above and silently disarm the largest
-/// weight in the formula.
-#[test]
-fn a_source_only_change_is_still_scored_as_asymmetric() {
-    let files = vec![(
-        40,
-        2,
-        "crates/code-intel-cli/src/change_risk/signals.rs".to_string(),
-    )];
-    let scored = score_subset(&files, &FileHistory::new(), 1_700_000_000);
-
-    let asymmetry = &scored.signals["testAsymmetry"];
-    assert_eq!(asymmetry["testFilesChanged"].as_u64(), Some(0));
-    assert_eq!(asymmetry["asymmetric"], true);
-    assert_eq!(asymmetry["subscore"], 1.0);
-}
-
-#[test]
-fn looks_like_fix_subject_matches_english_and_chinese_markers() {
-    assert!(looks_like_fix_subject("fix(gate): correct off-by-one"));
-    assert!(looks_like_fix_subject("Fix regression in parser"));
-    assert!(looks_like_fix_subject("修复解析器越界问题"));
-    assert!(looks_like_fix_subject("修正边界条件"));
-    assert!(!looks_like_fix_subject("feat: add change risk subcommand"));
-}
-
-#[test]
-fn tip_token_extracts_the_range_head() {
-    assert_eq!(tip_token("origin/main..HEAD"), "HEAD");
-    assert_eq!(tip_token("abc123...def456"), "def456");
-    assert_eq!(tip_token("origin/main.."), "HEAD");
-    assert_eq!(tip_token("abc123"), "abc123");
-}
 
 #[test]
 fn empty_diff_reports_a_warning_without_erroring() {
@@ -505,66 +300,4 @@ fn repo_flag_rejects_a_path_that_is_not_a_git_repository() {
     );
 
     std::fs::remove_dir_all(&not_a_repo).ok();
-}
-
-/// The workflow that consumes this module had no test at all, which is how
-/// issue #201 lived in it unnoticed: the gate judged `risk_percentile`, a rank
-/// against a rolling sample, so a pull request could turn red because its
-/// neighbours got smaller rather than because it got riskier — and ~10% of
-/// commits sit above the 90th percentile by construction, making the gate a
-/// quota no amount of repository health could clear.
-///
-/// This pins the contract, not the bytes: which field decides, that the
-/// percentile is reported but never gated, and that the threshold sits in the
-/// measured gap between the dense 66-69 cluster and the 90+ outliers. A silent
-/// revert to percentile gating, or a threshold quietly tuned down into the
-/// cluster to unblock something in flight, fails here.
-#[test]
-fn the_pr_gate_blocks_on_the_absolute_score_not_a_moving_rank() {
-    const WORKFLOW: &str = include_str!("../../../../.github/workflows/pr-gate.yml");
-
-    assert!(
-        WORKFLOW.contains("RISK_SCORE_BLOCK:"),
-        "pr-gate.yml must declare an absolute score threshold"
-    );
-    assert!(
-        !WORKFLOW.contains("RISK_PERCENTILE_BLOCK"),
-        "the percentile threshold is retired (#201); a reintroduced one gates on a moving denominator"
-    );
-
-    let threshold: f64 = WORKFLOW
-        .lines()
-        .find_map(|line| line.trim().strip_prefix("RISK_SCORE_BLOCK:"))
-        .map(|value| value.trim().trim_matches('"').to_string())
-        .expect("pr-gate.yml declares RISK_SCORE_BLOCK")
-        .parse()
-        .expect("RISK_SCORE_BLOCK is numeric");
-    assert!(
-        (70.0..=89.0).contains(&threshold),
-        "the threshold must stay inside the measured empty band between the 66-69 cluster and \
-         the 90+ outliers; re-derive the distribution before moving it outside: {threshold}"
-    );
-
-    // The gate reads `.score`; `risk_percentile` may only be narrated.
-    let gate_step = WORKFLOW
-        .split("- name: Evaluate gate")
-        .nth(1)
-        .expect("pr-gate.yml has an Evaluate gate step");
-    assert!(
-        gate_step.contains("'.score'") || gate_step.contains(".score >="),
-        "the gate step must decide on the absolute score"
-    );
-    assert!(
-        !gate_step.contains("risk_percentile"),
-        "the gate step must not read risk_percentile (#201)"
-    );
-
-    // `score` is a JSON number and `round2` can emit a fraction, which shell
-    // `-ge` rejects outright; under `bash -e` that aborts the job instead of
-    // deciding, so the comparison has to happen in jq.
-    assert!(
-        gate_step.contains("jq --argjson threshold"),
-        "the score comparison must run in jq, not shell -ge, so a fractional score cannot abort \
-         the job instead of blocking or passing"
-    );
 }
